requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
Brotli==1.1.0
cachetools==5.5.0
cassidy==0.1.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
Django==5.0.1
django-allauth==65.0.2
django-crum==0.7.9
django-dotenv==1.4.2
django-widget-tweaks==1.5.0
enumb==0.1.5
expo==0.1.2
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.67.0
h5py==3.12.1
httplib2==0.22.0
idna==3.10
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mimetype==0.1.5
ml-dtypes==0.4.1
namex==0.0.8
nibabel==5.3.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.2.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zopfli==0.2.3.post1